feat: QR 코드 UI/UX 변경
