+++ src/containers/qr-code/qr-code-action/qr-code-printer/qr-code-printer.constants.ts
