--- src/containers/qr-code/qr-code.type.ts
@@ -1,1 +1,0 @@
-export type QRCode = { value: string; name: string };
