+++ src/containers/qr-code/qr-code.types.ts
@@ -1,0 +1,8 @@
+export type QRCode = { value: string; name: string };
+
+export type SavedQRCodes = {
+  id: string;
+  date: string;
+  url: string;
+  title: string;
+}[];
